using long as score to prevent overflow

+++ src/ai/susi/mind/SusiIntent.java
@@ -322,7 +322,7 @@ public class SusiIntent {
      */
     public class Score {
 
-        public int score;
+        public long score;
         public String log;
         
         public Score(SusiLanguage userLanguage) {
